Refactor api to split category related permissions
