+++ src/accounts/api/views.py
@@ -6,7 +6,7 @@ from django.db.models import Q
 from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated, BasePermission
 
-from categories.models import Category
+from restapi.views import CategoryAPIViewMixin
 from accounts.models import User
 from accounts.api.serializers import UserSerializer
 
@@ -18,34 +18,11 @@ class DCPermissions(BasePermission):
         return request.user.has_perm(self.perm)
 
 
-class CategoryPermission(BasePermission):
-
-    def has_permission(self, request, view):
-        category = view.get_category()
-        if category is None:
-            return False
-        return request.user in category.users.all()
-
-
-class UserViewSet(viewsets.ReadOnlyModelViewSet):
+class UserViewSet(CategoryAPIViewMixin, viewsets.ReadOnlyModelViewSet):
     model = User
     serializer_class = UserSerializer
     paginate_by_param = 'page_limit'
-    permission_classes = (IsAuthenticated, DCPermissions, CategoryPermission)
-    _category = None
-
-    def get_category(self):
-        if self._category is None:
-            organisation_slug = self.kwargs.get('organisation')
-            category_slug = self.kwargs.get('category')
-            try:
-                self._category = Category.objects.get(
-                    organisation__slug=organisation_slug,
-                    category_template__slug=category_slug)
-            except Category.DoesNotExist:
-                self._category = None
-
-        return self._category
+    permission_classes = (IsAuthenticated, DCPermissions)
 
     def get_queryset(self):
         qs = User.objects.filter(categories=self.get_category())
